Move inspector controls to their own folder
